custom/requirements.txt
Flask==3.1.0
pandas==2.2.3
openai==1.55.0
python-dotenv==1.0.1